e2e: Run tests using Ansible and Vagrant
Create infrastructure that runs the e2e tests using Ansible and
Vagrant. The cri-rm e2e tests are used as a base but all the code
related to creating VM with a given distro is removed and replaced
by Vagrant and Ansible. This also means that govm based VM is no
longer used.