resmgr: clear any empty configuration from the cache.
Always reset an empty config (as opposed to no/nil config)
if found in an existing cache. Most policies fail to start
without some CPU reservation which an empty configuration
lacks. Resetting it will let the policy to start with its
fallback configuration which is fully controlled by the
user and should be usable by the policy.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>