e2e: Deploy configuration file as a configMap
test/e2e/policies.test-suite/topology-aware/n4c16/ has several test
cases and some with custom configuration file for the policy.
Previously, we tried to use configMap for the configuration file
but the issue was that the same configMap was used for all the test
cases. To overcome that issue, use kustomize to generate a configMap
based on the configuration file present in a test case. If a test
case doesn't have a custom configuration file, then kustomize uses
default configuration file (e.g. test/e2e/files/nri-resmgr.cfg for TA)
to generate the configMap.

Signed-off-by: Muyassarov, Feruzjon <feruzjon.muyassarov@intel.com>