e2e: Add/fix default nri-resmgr configuration file
The e2e tests will utilize the default nri-resmgr configuration
file if found in the policy directory and if no other config
file is specified by the test script.
For the balloons policy, we add the nri-plugin debugging as
that is useful to see in case there is an error between the plugin
and the container runtime.

Signed-off-by: Jukka Rissanen <jukka.rissanen@linux.intel.com>