resmgr: don't return an error when ignoring config from agent
Don't return an error when ignoring agent config with force config
enabled. Even if the config is incorrect, we should not return an
error and just continue because we are not on the fallback but in
force config (no matter if the config is correct or not we just ignore
it).

Signed-off-by: Muyassarov, Feruzjon <feruzjon.muyassarov@intel.com>