e2e: create fallback config file for configmap based TA tests
In case of readiness probe test, we want to use fallback config
based daemonSet in topology aWare plugin as we already do for
balloons plugin. The reason is, for health checking test, we supply
incorrect configuration in the form of configMap, which should be
picked up by the nri-rm. Without using fallback (i.e. force) nri-rm
simply ignores the configuration, which doesn't allow us to test
health-checking functionality.

Signed-off-by: Muyassarov, Feruzjon <feruzjon.muyassarov@intel.com>